core/exam/api/v1/urls/question_urls.py
from rest_framework.routers import DefaultRouter
from ..views import QuestionViewSet

router = DefaultRouter()

urlpatterns = [
]

router.register(r"questions", QuestionViewSet, basename="questions")

urlpatterns+=router.urls
